feat(v2): add option for changing duplicate path behavior
