+++ packages/docusaurus/src/server/plugins/index.ts
@@ -48,43 +48,6 @@ export function sortConfig(routeConfigs: RouteConfig[]): void {
   });
 }
 
-export function warnAboutOverridingRoutes(
-  pluginsRouteConfigs: RouteConfig[],
-): void {
-  // Accumulate all the routes by recursively exploring each RouteConfig
-  const routesAccumulator: string[] = [];
-  function getAllRoutes(routeConfigs: RouteConfig[]): string[] {
-    for (let i = 0; i < routeConfigs.length; i += 1) {
-      const routeConfig = routeConfigs[i];
-      routesAccumulator.push(routeConfig.path);
-      if (routeConfig.routes !== undefined) {
-        getAllRoutes(routeConfig.routes);
-      }
-    }
-    return routesAccumulator;
-  }
-
-  const allRoutes = getAllRoutes(pluginsRouteConfigs);
-
-  // Sort the allRoutes array in lexicographical order
-  // Then check if each route is equal to the next route
-  // If yes, one of these routes will be overridden so we warn the user
-  allRoutes.sort((a, b) => a.localeCompare(b));
-  for (let i = 0; i < allRoutes.length - 1; i += 1) {
-    if (allRoutes[i] === allRoutes[i + 1]) {
-      console.warn(
-        `${
-          chalk.yellow(`warning `) + chalk.bold.yellow(`Routes Override: `)
-        }Attempting to create page at "${
-          allRoutes[i]
-        }" but a page already exists at this path\n${chalk.bold.yellow(
-          `This could lead to non-deterministic routing behavior`,
-        )}`,
-      );
-    }
-  }
-}
-
 export async function loadPlugins({
   pluginConfigs,
   context,
@@ -171,8 +134,6 @@ export async function loadPlugins({
       });
     }),
   );
-
-  warnAboutOverridingRoutes(pluginsRouteConfigs);
 
   // 4. Plugin Lifecycle - routesLoaded.
   // Currently plugins run lifecycle methods in parallel and are not order-dependent.
